Cleanup dependencies, rename files
