--- fixr
@@ -1,61 +1,0 @@
-import scalaz._, schema._, Json._
-import monocle._
-
-object module extends JsonModule[JsonSchema.type] { 
-
-  val R = JsonSchema 
-  import R._
-final case class Person(name: String, role: Option[Role])
-sealed trait Role
-final case class User(active: Boolean, boss: Person) extends Role
-final case class Admin(rights: List[String])         extends Role
-
-
-
-  def user(pers: BareSchema[ Person]) = record(
-    "active" -*>: prim(JsonSchema.JsonBool) :*: "boss" -*>: self(pers),
-    Iso[(Boolean, Person), User]((User.apply _).tupled)(u => (u.active, u.boss))
-  )
-
-  val admin = record(
-    "rights" -*>: seq(prim(JsonSchema.JsonString)),
-    Iso[List[String], Admin](Admin.apply)(_.rights)
-  )
-
-  def role(pers: Schema_[Person]) = union(
-    "user" -+>: user(pers) :+:
-      "admin" -+>: admin,
-    Iso[User \/ Admin, Role] {
-      case -\/(u) => u
-      case \/-(a) => a
-    } {
-      case u @ User(_, _) => -\/(u)
-      case a @ Admin(_)   => \/-(a)
-    }
-  )
-
-  def person(pers: Schema_[Person]) = record(
-    "name" -*>: prim(JsonSchema.JsonString) :*:
-      "role" -*>: optional(
-      role(pers)
-    ),
-    Iso[(String, Option[Role]), Person]((Person.apply _).tupled)(p => (p.name, p.role))
-  )
-
-        implicit val primToEncoderNT = new (R.Prim ~> EncoderA) {
-          def apply[A](fa: R.Prim[A]): EncoderA[A] = { a =>
-            fa match {
-              case JsonNumber => a.toString
-              case JsonBool   => a.toString
-              case JsonString => s""""$a""""
-              case JsonNull   => "null"
-            }
-          }
-        }
-
-
-  val p = person(person(null))
-}
-import module._
-val burns = Person("Montgommery Burns", Some(Admin(List("billionaire", "evil mastermind"))))
-val homer = Person("Homer Simpson", Some(User(true, burns)))
